% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{alerts}
\alias{alerts}
\alias{alert_success}
\alias{alert_error}
\alias{alert_info}
\alias{alert_warning}
\title{Pre styled sweet alerts}
\usage{
alert_success(
  session = getDefaultReactiveDomain(),
  text = NULL,
  btn_labels = "Ok",
  btn_colors = "#3085d6",
  html = FALSE,
  closeOnClickOutside = TRUE,
  showCloseButton = FALSE,
  width = NULL,
  ...
)

alert_error(
  session = getDefaultReactiveDomain(),
  text = NULL,
  btn_labels = "Ok",
  btn_colors = "#3085d6",
  html = FALSE,
  closeOnClickOutside = TRUE,
  showCloseButton = FALSE,
  width = NULL,
  ...
)

alert_info(
  session = getDefaultReactiveDomain(),
  text = NULL,
  btn_labels = "Ok",
  btn_colors = "#3085d6",
  html = FALSE,
  closeOnClickOutside = TRUE,
  showCloseButton = FALSE,
  width = NULL,
  ...
)

alert_warning(
  session = getDefaultReactiveDomain(),
  text = NULL,
  btn_labels = "Ok",
  btn_colors = "#3085d6",
  html = FALSE,
  closeOnClickOutside = TRUE,
  showCloseButton = FALSE,
  width = NULL,
  ...
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to shinyServer.}

\item{text}{Text of the alert.}

\item{btn_labels}{Label(s) for button(s), can be of length 2,
in which case the alert will have two buttons. Use \code{NA} for no buttons.s}

\item{btn_colors}{Color(s) for the buttons.}

\item{html}{Does \code{text} contains HTML tags ?}

\item{closeOnClickOutside}{Decide whether the user should be able to dismiss
the modal by clicking outside of it, or not.}

\item{showCloseButton}{Show close button in top right corner of the modal.}

\item{width}{Width of the modal (in pixel).}

\item{...}{Other arguments passed to JavaScript method.}
}
\value{
HTML code for sweet alerts
}
\description{
These functions are convenient wrappers for \code{\link[shinyWidgets:sweetalert]{shinyWidgets::sendSweetAlert()}}.
}
\seealso{
\code{\link[shinyWidgets:sweetalert]{shinyWidgets::sendSweetAlert()}}
}
