% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-AppData.R
\name{AppData}
\alias{AppData}
\title{Get App data}
\description{
Get App data

Get App data
}
\details{
R6 class that allows to get the information needed for an User session of the app.
}
\section{Super classes}{
\code{\link[appReporteAvance:DBManager]{appReporteAvance::DBManager}} -> \code{\link[appReporteAvance:User]{appReporteAvance::User}} -> \code{\link[appReporteAvance:Organisation]{appReporteAvance::Organisation}} -> \code{\link[appReporteAvance:Group]{appReporteAvance::Group}} -> \code{appReporteAvance::Process} -> \code{\link[appReporteAvance:Task]{appReporteAvance::Task}} -> \code{AppData}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AppData-new}{\code{AppData$new()}}
\item \href{#method-AppData-clone}{\code{AppData$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="DBManager" data-id="db_execute_statement"><a href='../../appReporteAvance/html/DBManager.html#method-DBManager-db_execute_statement'><code>appReporteAvance::DBManager$db_execute_statement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="DBManager" data-id="db_get_query"><a href='../../appReporteAvance/html/DBManager.html#method-DBManager-db_get_query'><code>appReporteAvance::DBManager$db_get_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="DBManager" data-id="db_make_query"><a href='../../appReporteAvance/html/DBManager.html#method-DBManager-db_make_query'><code>appReporteAvance::DBManager$db_make_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="User" data-id="user_add"><a href='../../appReporteAvance/html/User.html#method-User-user_add'><code>appReporteAvance::User$user_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="User" data-id="user_delete"><a href='../../appReporteAvance/html/User.html#method-User-user_delete'><code>appReporteAvance::User$user_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="User" data-id="user_edit_names"><a href='../../appReporteAvance/html/User.html#method-User-user_edit_names'><code>appReporteAvance::User$user_edit_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_add"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_add'><code>appReporteAvance::Organisation$org_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_delete"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_delete'><code>appReporteAvance::Organisation$org_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_edit"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_edit'><code>appReporteAvance::Organisation$org_edit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_finalize"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_finalize'><code>appReporteAvance::Organisation$org_finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_initialize"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_initialize'><code>appReporteAvance::Organisation$org_initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_user_add"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_user_add'><code>appReporteAvance::Organisation$org_user_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_user_delete"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_user_delete'><code>appReporteAvance::Organisation$org_user_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_user_edit"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_user_edit'><code>appReporteAvance::Organisation$org_user_edit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_add"><a href='../../appReporteAvance/html/Group.html#method-Group-group_add'><code>appReporteAvance::Group$group_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_delete"><a href='../../appReporteAvance/html/Group.html#method-Group-group_delete'><code>appReporteAvance::Group$group_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_edit"><a href='../../appReporteAvance/html/Group.html#method-Group-group_edit'><code>appReporteAvance::Group$group_edit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_initialize"><a href='../../appReporteAvance/html/Group.html#method-Group-group_initialize'><code>appReporteAvance::Group$group_initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_select"><a href='../../appReporteAvance/html/Group.html#method-Group-group_select'><code>appReporteAvance::Group$group_select()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_set_as_favorite"><a href='../../appReporteAvance/html/Group.html#method-Group-group_set_as_favorite'><code>appReporteAvance::Group$group_set_as_favorite()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_unit_add"><a href='../../appReporteAvance/html/Group.html#method-Group-group_unit_add'><code>appReporteAvance::Group$group_unit_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_unit_delete"><a href='../../appReporteAvance/html/Group.html#method-Group-group_unit_delete'><code>appReporteAvance::Group$group_unit_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_unit_edit"><a href='../../appReporteAvance/html/Group.html#method-Group-group_unit_edit'><code>appReporteAvance::Group$group_unit_edit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_user_add"><a href='../../appReporteAvance/html/Group.html#method-Group-group_user_add'><code>appReporteAvance::Group$group_user_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_user_delete"><a href='../../appReporteAvance/html/Group.html#method-Group-group_user_delete'><code>appReporteAvance::Group$group_user_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_user_edit"><a href='../../appReporteAvance/html/Group.html#method-Group-group_user_edit'><code>appReporteAvance::Group$group_user_edit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Process" data-id="fetch_processes"><a href='../../appReporteAvance/html/Process.html#method-Process-fetch_processes'><code>appReporteAvance::Process$fetch_processes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Process" data-id="fetch_units"><a href='../../appReporteAvance/html/Process.html#method-Process-fetch_units'><code>appReporteAvance::Process$fetch_units()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Process" data-id="process_add"><a href='../../appReporteAvance/html/Process.html#method-Process-process_add'><code>appReporteAvance::Process$process_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Process" data-id="process_delete"><a href='../../appReporteAvance/html/Process.html#method-Process-process_delete'><code>appReporteAvance::Process$process_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Process" data-id="process_edit"><a href='../../appReporteAvance/html/Process.html#method-Process-process_edit'><code>appReporteAvance::Process$process_edit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Process" data-id="unit_add"><a href='../../appReporteAvance/html/Process.html#method-Process-unit_add'><code>appReporteAvance::Process$unit_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Process" data-id="unit_delete"><a href='../../appReporteAvance/html/Process.html#method-Process-unit_delete'><code>appReporteAvance::Process$unit_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Process" data-id="unit_edit"><a href='../../appReporteAvance/html/Process.html#method-Process-unit_edit'><code>appReporteAvance::Process$unit_edit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Task" data-id="progress_add"><a href='../../appReporteAvance/html/Task.html#method-Task-progress_add'><code>appReporteAvance::Task$progress_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Task" data-id="report_add"><a href='../../appReporteAvance/html/Task.html#method-Task-report_add'><code>appReporteAvance::Task$report_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Task" data-id="report_archive"><a href='../../appReporteAvance/html/Task.html#method-Task-report_archive'><code>appReporteAvance::Task$report_archive()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Task" data-id="report_delete"><a href='../../appReporteAvance/html/Task.html#method-Task-report_delete'><code>appReporteAvance::Task$report_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Task" data-id="task_add"><a href='../../appReporteAvance/html/Task.html#method-Task-task_add'><code>appReporteAvance::Task$task_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Task" data-id="task_archive"><a href='../../appReporteAvance/html/Task.html#method-Task-task_archive'><code>appReporteAvance::Task$task_archive()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Task" data-id="task_delete"><a href='../../appReporteAvance/html/Task.html#method-Task-task_delete'><code>appReporteAvance::Task$task_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Task" data-id="task_edit_metadata"><a href='../../appReporteAvance/html/Task.html#method-Task-task_edit_metadata'><code>appReporteAvance::Task$task_edit_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Task" data-id="task_get_history"><a href='../../appReporteAvance/html/Task.html#method-Task-task_get_history'><code>appReporteAvance::Task$task_get_history()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Task" data-id="task_report_progress"><a href='../../appReporteAvance/html/Task.html#method-Task-task_report_progress'><code>appReporteAvance::Task$task_report_progress()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppData-new"></a>}}
\if{latex}{\out{\hypertarget{method-AppData-new}{}}}
\subsection{Method \code{new()}}{
Start a session of app data based on an user email
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppData$new(email = Sys.getenv("REPORTES_EMAIL"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{The email the user started the session with.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AppData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
