% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btns.R
\name{btn_x}
\alias{btn_x}
\title{Pretty action buttons factory}
\usage{
btn_x(color, label)
}
\arguments{
\item{color}{Color of the button : \code{default}, \code{primary}, \code{warning},
\code{danger}, \code{success}, \code{royal}.}

\item{label}{The contents of the button, usually a text label.}
}
\value{
A function to generate action buttons with a particular styling
}
\description{
Pretty action buttons factory
}
