% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBManager.R
\name{DBManager}
\alias{DBManager}
\title{Clase R6 para manejar la base de datos}
\description{
Una base de datos puede ejecutar y obtener querys
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DBManager-clone}{\code{DBManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DBManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
