% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Group.R
\name{Group}
\alias{Group}
\title{Get Group data}
\description{
Get Group data

Get Group data
}
\details{
R6 class that allows to get the Group information.
}
\section{Super classes}{
\code{\link[appReporteAvance:DBManager]{appReporteAvance::DBManager}} -> \code{\link[appReporteAvance:User]{appReporteAvance::User}} -> \code{\link[appReporteAvance:Organisation]{appReporteAvance::Organisation}} -> \code{Group}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{group_selected}}{ID of the group to be used in the board, by default is the favorite group in group_users}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{groups}}{List containing the group affiliations of the User}

\item{\code{group_users}}{List containing the user list of the group The info is shown following the User's group role.}

\item{\code{group_units}}{List containing the group's measurement units. Older units are shown first.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Group-new}{\code{Group$new()}}
\item \href{#method-Group-group_initialize}{\code{Group$group_initialize()}}
\item \href{#method-Group-group_add}{\code{Group$group_add()}}
\item \href{#method-Group-group_delete}{\code{Group$group_delete()}}
\item \href{#method-Group-group_edit}{\code{Group$group_edit()}}
\item \href{#method-Group-group_user_add}{\code{Group$group_user_add()}}
\item \href{#method-Group-group_user_delete}{\code{Group$group_user_delete()}}
\item \href{#method-Group-group_user_edit}{\code{Group$group_user_edit()}}
\item \href{#method-Group-group_select}{\code{Group$group_select()}}
\item \href{#method-Group-group_set_as_favorite}{\code{Group$group_set_as_favorite()}}
\item \href{#method-Group-group_unit_add}{\code{Group$group_unit_add()}}
\item \href{#method-Group-group_unit_edit}{\code{Group$group_unit_edit()}}
\item \href{#method-Group-group_unit_delete}{\code{Group$group_unit_delete()}}
\item \href{#method-Group-clone}{\code{Group$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="DBManager" data-id="db_execute_statement"><a href='../../appReporteAvance/html/DBManager.html#method-DBManager-db_execute_statement'><code>appReporteAvance::DBManager$db_execute_statement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="DBManager" data-id="db_get_query"><a href='../../appReporteAvance/html/DBManager.html#method-DBManager-db_get_query'><code>appReporteAvance::DBManager$db_get_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="DBManager" data-id="db_make_query"><a href='../../appReporteAvance/html/DBManager.html#method-DBManager-db_make_query'><code>appReporteAvance::DBManager$db_make_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="User" data-id="user_add"><a href='../../appReporteAvance/html/User.html#method-User-user_add'><code>appReporteAvance::User$user_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="User" data-id="user_delete"><a href='../../appReporteAvance/html/User.html#method-User-user_delete'><code>appReporteAvance::User$user_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="User" data-id="user_edit_names"><a href='../../appReporteAvance/html/User.html#method-User-user_edit_names'><code>appReporteAvance::User$user_edit_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="fetch_orgs"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-fetch_orgs'><code>appReporteAvance::Organisation$fetch_orgs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_add"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_add'><code>appReporteAvance::Organisation$org_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_delete"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_delete'><code>appReporteAvance::Organisation$org_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_edit"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_edit'><code>appReporteAvance::Organisation$org_edit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_finalize"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_finalize'><code>appReporteAvance::Organisation$org_finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_initialize"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_initialize'><code>appReporteAvance::Organisation$org_initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_user_add"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_user_add'><code>appReporteAvance::Organisation$org_user_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_user_delete"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_user_delete'><code>appReporteAvance::Organisation$org_user_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_user_edit"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_user_edit'><code>appReporteAvance::Organisation$org_user_edit()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-new"></a>}}
\if{latex}{\out{\hypertarget{method-Group-new}{}}}
\subsection{Method \code{new()}}{
Start a Group based on an user email
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$new(email = Sys.getenv("REPORTES_EMAIL"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{The email the user started the session with.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-group_initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Group-group_initialize}{}}}
\subsection{Method \code{group_initialize()}}{
Initialize a group for a new user
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$group_initialize(org_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_id}}{The id of the organisation on which the statement will be executed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-group_add"></a>}}
\if{latex}{\out{\hypertarget{method-Group-group_add}{}}}
\subsection{Method \code{group_add()}}{
Add a group to the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$group_add(org_id, group_title, group_description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_id}}{The id of the organisation on which the statement will be executed}

\item{\code{group_title}}{The new title of the group}

\item{\code{group_description}}{The new description of the group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-group_delete"></a>}}
\if{latex}{\out{\hypertarget{method-Group-group_delete}{}}}
\subsection{Method \code{group_delete()}}{
Remove a group from the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$group_delete(group_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_id}}{The id of the group on which the statement will be executed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-group_edit"></a>}}
\if{latex}{\out{\hypertarget{method-Group-group_edit}{}}}
\subsection{Method \code{group_edit()}}{
Edit group metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$group_edit(group_id, group_title, group_description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_id}}{The id of the group on which the statement will be executed}

\item{\code{group_title}}{The new title of the group}

\item{\code{group_description}}{The new description of the group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-group_user_add"></a>}}
\if{latex}{\out{\hypertarget{method-Group-group_user_add}{}}}
\subsection{Method \code{group_user_add()}}{
Add an user to a group
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$group_user_add(
  group_id,
  user_id,
  user_color = "white",
  group_role = "user"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_id}}{The id of the group on which the statement will be executed}

\item{\code{user_id}}{The id of the user on which the statement will be executed}

\item{\code{user_color}}{The color of the user's cards}

\item{\code{group_role}}{The role for the user in the group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-group_user_delete"></a>}}
\if{latex}{\out{\hypertarget{method-Group-group_user_delete}{}}}
\subsection{Method \code{group_user_delete()}}{
Delete an user from a group
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$group_user_delete(group_id, user_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_id}}{The id of the group on which the statement will be executed}

\item{\code{user_id}}{The id of the user on which the statement will be executed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-group_user_edit"></a>}}
\if{latex}{\out{\hypertarget{method-Group-group_user_edit}{}}}
\subsection{Method \code{group_user_edit()}}{
Edit the role of a user inside a group and related information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$group_user_edit(group_id, user_id, user_color, group_role)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_id}}{The id of the group on which the statement will be executed}

\item{\code{user_id}}{The id of the user on which the statement will be executed}

\item{\code{user_color}}{The color of the user's cards}

\item{\code{group_role}}{The role for the user in the group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-group_select"></a>}}
\if{latex}{\out{\hypertarget{method-Group-group_select}{}}}
\subsection{Method \code{group_select()}}{
Select a group for use in the board
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$group_select(group_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_id}}{The id of the group on which the statement will be executed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-group_set_as_favorite"></a>}}
\if{latex}{\out{\hypertarget{method-Group-group_set_as_favorite}{}}}
\subsection{Method \code{group_set_as_favorite()}}{
Set selected group as favorite in the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$group_set_as_favorite()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-group_unit_add"></a>}}
\if{latex}{\out{\hypertarget{method-Group-group_unit_add}{}}}
\subsection{Method \code{group_unit_add()}}{
Add a measurement unit for a group
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$group_unit_add(
  unit_title,
  unit_description = "",
  unit_type,
  unit_icon = "file"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{unit_title}}{The title of a group's measurement unit}

\item{\code{unit_description}}{The description of a group's measurement unit}

\item{\code{unit_type}}{The type of a group's measurement unit. One of ("report", "task")}

\item{\code{unit_icon}}{The icon of a group's measurement unit. Should be the compatible with fontawesome::fa().}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-group_unit_edit"></a>}}
\if{latex}{\out{\hypertarget{method-Group-group_unit_edit}{}}}
\subsection{Method \code{group_unit_edit()}}{
Edit a measurement unit from a group
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$group_unit_edit(
  unit_id,
  unit_title,
  unit_description,
  unit_type,
  unit_icon
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{unit_id}}{The id of a group's measurement unit}

\item{\code{unit_title}}{The title of a group's measurement unit}

\item{\code{unit_description}}{The description of a group's measurement unit}

\item{\code{unit_type}}{The type of a group's measurement unit. One of ("report", "task")}

\item{\code{unit_icon}}{The icon of a group's measurement unit. Should be the compatible with fontawesome::fa().}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-group_unit_delete"></a>}}
\if{latex}{\out{\hypertarget{method-Group-group_unit_delete}{}}}
\subsection{Method \code{group_unit_delete()}}{
Delete a measurement unit from a group
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$group_unit_delete(unit_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{unit_id}}{The id of a group's measurement unit}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Group-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
