% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_icon_picker.R
\name{iconPicker_dep}
\alias{iconPicker_dep}
\title{Icon Picker Dependency}
\usage{
iconPicker_dep()
}
\value{
a html dependency named "iconPicker".
}
\description{
Create a dependency for the icon picker element.
}
