% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_alerts.R
\name{mk_alert}
\alias{mk_alert}
\title{Sweet alerts factory}
\usage{
mk_alert(title = "Title", type = NULL)
}
\arguments{
\item{title}{Title of the alert.}

\item{type}{Type of the alert : info, success, warning or error.}
}
\value{
A function to generate a sweet alert with a particular title and type.
}
\description{
Sweet alerts factory
}
