% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Orgs.R
\name{Organisation}
\alias{Organisation}
\title{Get Organisation data}
\description{
Get Organisation data

Get Organisation data
}
\details{
R6 class that allows to get the Organisation information.
}
\section{Super classes}{
\code{\link[appReporteAvance:DBManager]{appReporteAvance::DBManager}} -> \code{\link[appReporteAvance:User]{appReporteAvance::User}} -> \code{Organisation}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{orgs}}{List containing the organisation affiliations of the User}

\item{\code{org_users}}{List containing the user list of the organisation. The info is shown following the User's organisation role.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Organisation-new}{\code{Organisation$new()}}
\item \href{#method-Organisation-org_initialize}{\code{Organisation$org_initialize()}}
\item \href{#method-Organisation-org_add}{\code{Organisation$org_add()}}
\item \href{#method-Organisation-org_delete}{\code{Organisation$org_delete()}}
\item \href{#method-Organisation-org_edit}{\code{Organisation$org_edit()}}
\item \href{#method-Organisation-org_user_add}{\code{Organisation$org_user_add()}}
\item \href{#method-Organisation-org_user_delete}{\code{Organisation$org_user_delete()}}
\item \href{#method-Organisation-org_user_edit}{\code{Organisation$org_user_edit()}}
\item \href{#method-Organisation-org_finalize}{\code{Organisation$org_finalize()}}
\item \href{#method-Organisation-clone}{\code{Organisation$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="DBManager" data-id="db_execute_statement"><a href='../../appReporteAvance/html/DBManager.html#method-DBManager-db_execute_statement'><code>appReporteAvance::DBManager$db_execute_statement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="DBManager" data-id="db_get_query"><a href='../../appReporteAvance/html/DBManager.html#method-DBManager-db_get_query'><code>appReporteAvance::DBManager$db_get_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="DBManager" data-id="db_make_query"><a href='../../appReporteAvance/html/DBManager.html#method-DBManager-db_make_query'><code>appReporteAvance::DBManager$db_make_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="User" data-id="user_add"><a href='../../appReporteAvance/html/User.html#method-User-user_add'><code>appReporteAvance::User$user_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="User" data-id="user_delete"><a href='../../appReporteAvance/html/User.html#method-User-user_delete'><code>appReporteAvance::User$user_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="User" data-id="user_edit_names"><a href='../../appReporteAvance/html/User.html#method-User-user_edit_names'><code>appReporteAvance::User$user_edit_names()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Organisation-new"></a>}}
\if{latex}{\out{\hypertarget{method-Organisation-new}{}}}
\subsection{Method \code{new()}}{
Start an Organisation based on an user email
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organisation$new(email)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{The email the user started the session with.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Organisation-org_initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Organisation-org_initialize}{}}}
\subsection{Method \code{org_initialize()}}{
Initialize an organisation for a new user
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organisation$org_initialize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Organisation-org_add"></a>}}
\if{latex}{\out{\hypertarget{method-Organisation-org_add}{}}}
\subsection{Method \code{org_add()}}{
Add a new organisation to the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organisation$org_add(org_title, org_description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_title}}{The new title of the organisation}

\item{\code{org_description}}{The new description of the organisation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Organisation-org_delete"></a>}}
\if{latex}{\out{\hypertarget{method-Organisation-org_delete}{}}}
\subsection{Method \code{org_delete()}}{
Delete an organisation from the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organisation$org_delete(org_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_id}}{The id of the organisation on which the statement will be executed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Organisation-org_edit"></a>}}
\if{latex}{\out{\hypertarget{method-Organisation-org_edit}{}}}
\subsection{Method \code{org_edit()}}{
Edit Organisation metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organisation$org_edit(org_id, org_title, org_description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_id}}{The id of the organisation on which the statement will be executed}

\item{\code{org_title}}{The new title of the organisation}

\item{\code{org_description}}{The new description of the organisation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Organisation-org_user_add"></a>}}
\if{latex}{\out{\hypertarget{method-Organisation-org_user_add}{}}}
\subsection{Method \code{org_user_add()}}{
Add an user to an organisation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organisation$org_user_add(org_id, user_id, org_role)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_id}}{The id of the organisation on which the statement will be executed}

\item{\code{user_id}}{The id of the user on which the statement will be executed}

\item{\code{org_role}}{The role for the user in the organisation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Organisation-org_user_delete"></a>}}
\if{latex}{\out{\hypertarget{method-Organisation-org_user_delete}{}}}
\subsection{Method \code{org_user_delete()}}{
Delete an user from an organisation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organisation$org_user_delete(org_id, user_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_id}}{The id of the organisation on which the statement will be executed}

\item{\code{user_id}}{The id of the user on which the statement will be executed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Organisation-org_user_edit"></a>}}
\if{latex}{\out{\hypertarget{method-Organisation-org_user_edit}{}}}
\subsection{Method \code{org_user_edit()}}{
Edit the role of a user inside an organisation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organisation$org_user_edit(org_id, user_id, org_role)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_id}}{The id of the organisation on which the statement will be executed}

\item{\code{user_id}}{The id of the user on which the statement will be executed}

\item{\code{org_role}}{The role for the user in the organisation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Organisation-org_finalize"></a>}}
\if{latex}{\out{\hypertarget{method-Organisation-org_finalize}{}}}
\subsection{Method \code{org_finalize()}}{
Remove the existence of an organisation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organisation$org_finalize(org_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_id}}{The id of the organisation on which the statement will be executed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Organisation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Organisation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organisation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
