% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Process.R
\name{Process}
\alias{Process}
\title{Process Class Definition}
\description{
This is a class definition for the Process class. It inherits from the Group class.
}
\section{Super classes}{
\code{\link[appReporteAvance:DBManager]{appReporteAvance::DBManager}} -> \code{\link[appReporteAvance:User]{appReporteAvance::User}} -> \code{\link[appReporteAvance:Organisation]{appReporteAvance::Organisation}} -> \code{\link[appReporteAvance:Group]{appReporteAvance::Group}} -> \code{Process}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Process-new}{\code{Process$new()}}
\item \href{#method-Process-fetch_processes}{\code{Process$fetch_processes()}}
\item \href{#method-Process-fetch_units}{\code{Process$fetch_units()}}
\item \href{#method-Process-process_add}{\code{Process$process_add()}}
\item \href{#method-Process-process_delete}{\code{Process$process_delete()}}
\item \href{#method-Process-process_edit}{\code{Process$process_edit()}}
\item \href{#method-Process-unit_add}{\code{Process$unit_add()}}
\item \href{#method-Process-unit_edit}{\code{Process$unit_edit()}}
\item \href{#method-Process-unit_delete}{\code{Process$unit_delete()}}
\item \href{#method-Process-clone}{\code{Process$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="DBManager" data-id="db_execute_statement"><a href='../../appReporteAvance/html/DBManager.html#method-DBManager-db_execute_statement'><code>appReporteAvance::DBManager$db_execute_statement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="DBManager" data-id="db_get_query"><a href='../../appReporteAvance/html/DBManager.html#method-DBManager-db_get_query'><code>appReporteAvance::DBManager$db_get_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="DBManager" data-id="db_make_query"><a href='../../appReporteAvance/html/DBManager.html#method-DBManager-db_make_query'><code>appReporteAvance::DBManager$db_make_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="User" data-id="user_add"><a href='../../appReporteAvance/html/User.html#method-User-user_add'><code>appReporteAvance::User$user_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="User" data-id="user_delete"><a href='../../appReporteAvance/html/User.html#method-User-user_delete'><code>appReporteAvance::User$user_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="User" data-id="user_edit_names"><a href='../../appReporteAvance/html/User.html#method-User-user_edit_names'><code>appReporteAvance::User$user_edit_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="fetch_orgs"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-fetch_orgs'><code>appReporteAvance::Organisation$fetch_orgs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_add"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_add'><code>appReporteAvance::Organisation$org_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_delete"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_delete'><code>appReporteAvance::Organisation$org_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_edit"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_edit'><code>appReporteAvance::Organisation$org_edit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_finalize"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_finalize'><code>appReporteAvance::Organisation$org_finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_initialize"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_initialize'><code>appReporteAvance::Organisation$org_initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_user_add"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_user_add'><code>appReporteAvance::Organisation$org_user_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_user_delete"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_user_delete'><code>appReporteAvance::Organisation$org_user_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_user_edit"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_user_edit'><code>appReporteAvance::Organisation$org_user_edit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_add"><a href='../../appReporteAvance/html/Group.html#method-Group-group_add'><code>appReporteAvance::Group$group_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_delete"><a href='../../appReporteAvance/html/Group.html#method-Group-group_delete'><code>appReporteAvance::Group$group_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_edit"><a href='../../appReporteAvance/html/Group.html#method-Group-group_edit'><code>appReporteAvance::Group$group_edit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_initialize"><a href='../../appReporteAvance/html/Group.html#method-Group-group_initialize'><code>appReporteAvance::Group$group_initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_select"><a href='../../appReporteAvance/html/Group.html#method-Group-group_select'><code>appReporteAvance::Group$group_select()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_set_as_favorite"><a href='../../appReporteAvance/html/Group.html#method-Group-group_set_as_favorite'><code>appReporteAvance::Group$group_set_as_favorite()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_unit_add"><a href='../../appReporteAvance/html/Group.html#method-Group-group_unit_add'><code>appReporteAvance::Group$group_unit_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_unit_delete"><a href='../../appReporteAvance/html/Group.html#method-Group-group_unit_delete'><code>appReporteAvance::Group$group_unit_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_unit_edit"><a href='../../appReporteAvance/html/Group.html#method-Group-group_unit_edit'><code>appReporteAvance::Group$group_unit_edit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_user_add"><a href='../../appReporteAvance/html/Group.html#method-Group-group_user_add'><code>appReporteAvance::Group$group_user_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_user_delete"><a href='../../appReporteAvance/html/Group.html#method-Group-group_user_delete'><code>appReporteAvance::Group$group_user_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_user_edit"><a href='../../appReporteAvance/html/Group.html#method-Group-group_user_edit'><code>appReporteAvance::Group$group_user_edit()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Process-new"></a>}}
\if{latex}{\out{\hypertarget{method-Process-new}{}}}
\subsection{Method \code{new()}}{
This function initializes a Process object.
It calls the initialize() method of the Group class to inherit its properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$new(email)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{character string with user's email address}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Process-fetch_processes"></a>}}
\if{latex}{\out{\hypertarget{method-Process-fetch_processes}{}}}
\subsection{Method \code{fetch_processes()}}{
This function fetches all the processes that belong to a group using the group_selected property.
It uses the glue and DBI libraries to execute an SQL query on the database connection stored in the self$con property.
The results are transformed using purrr::pmap(list).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$fetch_processes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a list of processed retrieved from the database
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Process-fetch_units"></a>}}
\if{latex}{\out{\hypertarget{method-Process-fetch_units}{}}}
\subsection{Method \code{fetch_units()}}{
This function fetches all the units that belong to a process identified by its ID using the glue and DBI libraries to execute an SQL query on the database connection stored in the self$con property.
The results are transformed using purrr::pmap(list).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$fetch_units(process_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{process_id}}{integer with the ID of the process to retrieve units from}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list of units retrieved from the database
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Process-process_add"></a>}}
\if{latex}{\out{\hypertarget{method-Process-process_add}{}}}
\subsection{Method \code{process_add()}}{
This function adds a process to the database using SQL queries executed through the glue and DBI libraries.
It uses the ids library to generate random IDs for the new process and stores it in the ID column of the processes table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$process_add(title, description = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{character string with the title of the process to add}

\item{\code{description}}{character string with the description of the process (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the ID of the newly added process
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Process-process_delete"></a>}}
\if{latex}{\out{\hypertarget{method-Process-process_delete}{}}}
\subsection{Method \code{process_delete()}}{
This function deletes a process from the database using an SQL query executed through the glue and DBI libraries.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$process_delete(process_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{process_id}}{integer with the ID of the process to delete}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Process-process_edit"></a>}}
\if{latex}{\out{\hypertarget{method-Process-process_edit}{}}}
\subsection{Method \code{process_edit()}}{
This function edits the title and description of a process identified by its ID in the database.
It uses SQL queries to update the title and description columns of the processes table executed through the glue and DBI libraries.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$process_edit(process_id, title, description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{process_id}}{integer with the ID of the process to edit}

\item{\code{title}}{character string with the new title for the process}

\item{\code{description}}{character string with the new description for the process}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Process-unit_add"></a>}}
\if{latex}{\out{\hypertarget{method-Process-unit_add}{}}}
\subsection{Method \code{unit_add()}}{
This function adds a measurement unit for a process identified by its ID to the database using SQL queries executed through the glue and DBI libraries.
It uses the ids library to generate random IDs for the new unit and stores it in the ID column of the units table.
It also checks if the unit type is either "report" or "task", and sets the "type" column of the units table accordingly.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$unit_add(
  process_id,
  unit_title,
  unit_description = "",
  unit_type,
  unit_icon = "file"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{process_id}}{integer with the ID of the process to add a unit to}

\item{\code{unit_title}}{character string with the title of the unit to add}

\item{\code{unit_description}}{character string with the description of the unit (optional, defaults to "")}

\item{\code{unit_type}}{character string with the type of the unit, must be either "report" or "task"}

\item{\code{unit_icon}}{character string with the icon for the unit (optional, defaults to "file")}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the ID of the newly added unit
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Process-unit_edit"></a>}}
\if{latex}{\out{\hypertarget{method-Process-unit_edit}{}}}
\subsection{Method \code{unit_edit()}}{
This function edits the title, description, icon and/or type of a measurement unit identified by its ID in the database.
It uses SQL queries to update the title, description, icon and/or type columns of the units table executed through the glue and DBI libraries.
It also checks if the unit type is either "report" or "task", and sets the "type" column of the units table accordingly.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$unit_edit(unit_id, unit_title, unit_description, unit_type, unit_icon)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{unit_id}}{integer with the ID of the unit to be edited}

\item{\code{unit_title}}{character string with the new title for the unit}

\item{\code{unit_description}}{character string with the new description for the unit}

\item{\code{unit_type}}{character string with the new type of the unit, must be either "report" or "task"}

\item{\code{unit_icon}}{character string with the new icon for the unit}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Process-unit_delete"></a>}}
\if{latex}{\out{\hypertarget{method-Process-unit_delete}{}}}
\subsection{Method \code{unit_delete()}}{
This function deletes a measurement unit identified by its ID from the database using an SQL query executed through the glue and DBI libraries.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$unit_delete(unit_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{unit_id}}{integer with the ID of the unit to delete}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Process-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Process-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
