% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_icon_picker.R
\name{reportes_icon_names}
\alias{reportes_icon_names}
\title{Get Fontawesome Icons}
\usage{
reportes_icon_names()
}
\value{
a character vector with the names of the available icons.
}
\description{
Get the available solid style icons in the fontawesome package.
}
