% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Task.R
\name{Task}
\alias{Task}
\title{Get Group data}
\description{
Get Group data

Get Group data
}
\details{
R6 class that allows to get the Group information.
}
\section{Super classes}{
\code{\link[appReporteAvance:DBManager]{appReporteAvance::DBManager}} -> \code{\link[appReporteAvance:User]{appReporteAvance::User}} -> \code{\link[appReporteAvance:Organisation]{appReporteAvance::Organisation}} -> \code{\link[appReporteAvance:Group]{appReporteAvance::Group}} -> \code{\link[appReporteAvance:Process]{appReporteAvance::Process}} -> \code{Task}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{tasks}}{List containing the tasks an User can interact with}

\item{\code{reports}}{List containing the reports an User can interact with}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Task-new}{\code{Task$new()}}
\item \href{#method-Task-task_add}{\code{Task$task_add()}}
\item \href{#method-Task-task_delete}{\code{Task$task_delete()}}
\item \href{#method-Task-task_edit}{\code{Task$task_edit()}}
\item \href{#method-Task-task_report_progress}{\code{Task$task_report_progress()}}
\item \href{#method-Task-task_archive}{\code{Task$task_archive()}}
\item \href{#method-Task-progress_add}{\code{Task$progress_add()}}
\item \href{#method-Task-task_get_history}{\code{Task$task_get_history()}}
\item \href{#method-Task-report_add}{\code{Task$report_add()}}
\item \href{#method-Task-report_delete}{\code{Task$report_delete()}}
\item \href{#method-Task-report_archive}{\code{Task$report_archive()}}
\item \href{#method-Task-clone}{\code{Task$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="DBManager" data-id="db_execute_statement"><a href='../../appReporteAvance/html/DBManager.html#method-DBManager-db_execute_statement'><code>appReporteAvance::DBManager$db_execute_statement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="DBManager" data-id="db_get_query"><a href='../../appReporteAvance/html/DBManager.html#method-DBManager-db_get_query'><code>appReporteAvance::DBManager$db_get_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="DBManager" data-id="db_make_query"><a href='../../appReporteAvance/html/DBManager.html#method-DBManager-db_make_query'><code>appReporteAvance::DBManager$db_make_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="User" data-id="user_add"><a href='../../appReporteAvance/html/User.html#method-User-user_add'><code>appReporteAvance::User$user_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="User" data-id="user_delete"><a href='../../appReporteAvance/html/User.html#method-User-user_delete'><code>appReporteAvance::User$user_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="User" data-id="user_edit_names"><a href='../../appReporteAvance/html/User.html#method-User-user_edit_names'><code>appReporteAvance::User$user_edit_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_add"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_add'><code>appReporteAvance::Organisation$org_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_delete"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_delete'><code>appReporteAvance::Organisation$org_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_edit"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_edit'><code>appReporteAvance::Organisation$org_edit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_finalize"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_finalize'><code>appReporteAvance::Organisation$org_finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_initialize"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_initialize'><code>appReporteAvance::Organisation$org_initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_user_add"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_user_add'><code>appReporteAvance::Organisation$org_user_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_user_delete"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_user_delete'><code>appReporteAvance::Organisation$org_user_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Organisation" data-id="org_user_edit"><a href='../../appReporteAvance/html/Organisation.html#method-Organisation-org_user_edit'><code>appReporteAvance::Organisation$org_user_edit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_add"><a href='../../appReporteAvance/html/Group.html#method-Group-group_add'><code>appReporteAvance::Group$group_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_delete"><a href='../../appReporteAvance/html/Group.html#method-Group-group_delete'><code>appReporteAvance::Group$group_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_edit"><a href='../../appReporteAvance/html/Group.html#method-Group-group_edit'><code>appReporteAvance::Group$group_edit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_initialize"><a href='../../appReporteAvance/html/Group.html#method-Group-group_initialize'><code>appReporteAvance::Group$group_initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_select"><a href='../../appReporteAvance/html/Group.html#method-Group-group_select'><code>appReporteAvance::Group$group_select()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_set_as_favorite"><a href='../../appReporteAvance/html/Group.html#method-Group-group_set_as_favorite'><code>appReporteAvance::Group$group_set_as_favorite()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_unit_add"><a href='../../appReporteAvance/html/Group.html#method-Group-group_unit_add'><code>appReporteAvance::Group$group_unit_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_unit_delete"><a href='../../appReporteAvance/html/Group.html#method-Group-group_unit_delete'><code>appReporteAvance::Group$group_unit_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_unit_edit"><a href='../../appReporteAvance/html/Group.html#method-Group-group_unit_edit'><code>appReporteAvance::Group$group_unit_edit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_user_add"><a href='../../appReporteAvance/html/Group.html#method-Group-group_user_add'><code>appReporteAvance::Group$group_user_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_user_delete"><a href='../../appReporteAvance/html/Group.html#method-Group-group_user_delete'><code>appReporteAvance::Group$group_user_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Group" data-id="group_user_edit"><a href='../../appReporteAvance/html/Group.html#method-Group-group_user_edit'><code>appReporteAvance::Group$group_user_edit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Process" data-id="fetch_processes"><a href='../../appReporteAvance/html/Process.html#method-Process-fetch_processes'><code>appReporteAvance::Process$fetch_processes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Process" data-id="fetch_units"><a href='../../appReporteAvance/html/Process.html#method-Process-fetch_units'><code>appReporteAvance::Process$fetch_units()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Process" data-id="process_add"><a href='../../appReporteAvance/html/Process.html#method-Process-process_add'><code>appReporteAvance::Process$process_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Process" data-id="process_delete"><a href='../../appReporteAvance/html/Process.html#method-Process-process_delete'><code>appReporteAvance::Process$process_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Process" data-id="process_edit"><a href='../../appReporteAvance/html/Process.html#method-Process-process_edit'><code>appReporteAvance::Process$process_edit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Process" data-id="unit_add"><a href='../../appReporteAvance/html/Process.html#method-Process-unit_add'><code>appReporteAvance::Process$unit_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Process" data-id="unit_delete"><a href='../../appReporteAvance/html/Process.html#method-Process-unit_delete'><code>appReporteAvance::Process$unit_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="Process" data-id="unit_edit"><a href='../../appReporteAvance/html/Process.html#method-Process-unit_edit'><code>appReporteAvance::Process$unit_edit()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Task-new"></a>}}
\if{latex}{\out{\hypertarget{method-Task-new}{}}}
\subsection{Method \code{new()}}{
Start a Task based on an user email
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Task$new(email)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{The email the user started the session with.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Task-task_add"></a>}}
\if{latex}{\out{\hypertarget{method-Task-task_add}{}}}
\subsection{Method \code{task_add()}}{
Add a new task for an User and report it as initial progress.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Task$task_add(
  group_id,
  task_title,
  task_description,
  assignee,
  time_due,
  output_unit
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_id}}{The id of the group on which the statement will be executed}

\item{\code{task_title}}{Title for the task}

\item{\code{task_description}}{Long description of the task}

\item{\code{assignee}}{The id of the user responsible for the task}

\item{\code{time_due}}{Deadline for the task completion. Datetime}

\item{\code{output_unit}}{Unit of measurement of the task output}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Task-task_delete"></a>}}
\if{latex}{\out{\hypertarget{method-Task-task_delete}{}}}
\subsection{Method \code{task_delete()}}{
Delete an user task
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Task$task_delete(task_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task_id}}{The id of the task on which the statement will be executed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Task-task_edit"></a>}}
\if{latex}{\out{\hypertarget{method-Task-task_edit}{}}}
\subsection{Method \code{task_edit()}}{
Edit a task metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Task$task_edit(
  task_id,
  task_title = NULL,
  task_description = NULL,
  time_due = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task_id}}{The id of the task on which the statement will be executed}

\item{\code{task_title}}{Title for the task}

\item{\code{task_description}}{Long description of the task}

\item{\code{time_due}}{Deadline for the task completion. Datetime}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Task-task_report_progress"></a>}}
\if{latex}{\out{\hypertarget{method-Task-task_report_progress}{}}}
\subsection{Method \code{task_report_progress()}}{
Report progress on an assigned task
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Task$task_report_progress(task_id, status_current, details)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task_id}}{The id of the task on which the statement will be executed}

\item{\code{status_current}}{Current status of the specified task}

\item{\code{details}}{Explanation of the progress made}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Task-task_archive"></a>}}
\if{latex}{\out{\hypertarget{method-Task-task_archive}{}}}
\subsection{Method \code{task_archive()}}{
Archive a task
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Task$task_archive(task_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task_id}}{The id of the task on which the statement will be executed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Task-progress_add"></a>}}
\if{latex}{\out{\hypertarget{method-Task-progress_add}{}}}
\subsection{Method \code{progress_add()}}{
Insert progress info on some task
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Task$progress_add(task_id, status, details)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task_id}}{The id of the task on which the statement will be executed}

\item{\code{status}}{The status of the task once the new progress is added}

\item{\code{details}}{Explanation of the progress made}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Task-task_get_history"></a>}}
\if{latex}{\out{\hypertarget{method-Task-task_get_history}{}}}
\subsection{Method \code{task_get_history()}}{
Get a task's progression history
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Task$task_get_history(task_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task_id}}{The id of the task on which the statement will be executed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Task-report_add"></a>}}
\if{latex}{\out{\hypertarget{method-Task-report_add}{}}}
\subsection{Method \code{report_add()}}{
Add a report and the quantities it has contributed in the specified units
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Task$report_add(report_title, details, units, quantities)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{report_title}}{The title of the report}

\item{\code{details}}{Explanation of the progress made}

\item{\code{units}}{Units of the report. Must have the same size as quantities.}

\item{\code{quantities}}{Quantities of the report. Must have the same size as units.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Task-report_delete"></a>}}
\if{latex}{\out{\hypertarget{method-Task-report_delete}{}}}
\subsection{Method \code{report_delete()}}{
Delete a report and its contributions from the database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Task$report_delete(report_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{report_id}}{The id of the report on which the statement will be executed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Task-report_archive"></a>}}
\if{latex}{\out{\hypertarget{method-Task-report_archive}{}}}
\subsection{Method \code{report_archive()}}{
Archive a report and its contributions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Task$report_archive(report_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{report_id}}{The id of the report on which the statement will be executed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Task-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Task-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Task$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
