% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textInputPro.R
\name{textInputPro}
\alias{textInputPro}
\title{Create a text input with enhanced parameters}
\usage{
textInputPro(
  inputId,
  label,
  value = "",
  width = NULL,
  placeholder = NULL,
  maxlength = NULL,
  maxlengthCounter = FALSE,
  readonly = FALSE
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:validateCssUnit]{validateCssUnit()}}.}

\item{placeholder}{A character string giving the user a hint as to what can
be entered into the control. Internet Explorer 8 and 9 do not support this
option.}

\item{maxlength}{the maximum length of the input}

\item{maxlengthCounter}{whether or not to show a character counter}
}
\value{
Returns a text input with the specified parameters.
}
\description{
Create a text input with enhanced parameters
}
