% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-User.R
\name{User}
\alias{User}
\title{Get user information}
\description{
Get user information

Get user information
}
\details{
R6 class that allows to get the user information.
}
\section{Super class}{
\code{\link[appReporteAvance:DBManager]{appReporteAvance::DBManager}} -> \code{User}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{user}}{List containing the user info.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-User-new}{\code{User$new()}}
\item \href{#method-User-user_add}{\code{User$user_add()}}
\item \href{#method-User-user_delete}{\code{User$user_delete()}}
\item \href{#method-User-user_edit_names}{\code{User$user_edit_names()}}
\item \href{#method-User-clone}{\code{User$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="DBManager" data-id="db_execute_statement"><a href='../../appReporteAvance/html/DBManager.html#method-DBManager-db_execute_statement'><code>appReporteAvance::DBManager$db_execute_statement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="DBManager" data-id="db_get_query"><a href='../../appReporteAvance/html/DBManager.html#method-DBManager-db_get_query'><code>appReporteAvance::DBManager$db_get_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="appReporteAvance" data-topic="DBManager" data-id="db_make_query"><a href='../../appReporteAvance/html/DBManager.html#method-DBManager-db_make_query'><code>appReporteAvance::DBManager$db_make_query()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-User-new"></a>}}
\if{latex}{\out{\hypertarget{method-User-new}{}}}
\subsection{Method \code{new()}}{
Start User
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$new(email)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{The email the user started the session with.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-User-user_add"></a>}}
\if{latex}{\out{\hypertarget{method-User-user_add}{}}}
\subsection{Method \code{user_add()}}{
Add a new user to the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$user_add(name, last_name, email)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name to be inserted as user metadata.}

\item{\code{last_name}}{Last name to be inserted as user metadata.}

\item{\code{email}}{The email the user started the session with.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-User-user_delete"></a>}}
\if{latex}{\out{\hypertarget{method-User-user_delete}{}}}
\subsection{Method \code{user_delete()}}{
Delete a user from the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$user_delete(user_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_id}}{ID of the user to edit or delete.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-User-user_edit_names"></a>}}
\if{latex}{\out{\hypertarget{method-User-user_edit_names}{}}}
\subsection{Method \code{user_edit_names()}}{
Change the name and last name of the User
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$user_edit_names(user_id, name, last_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_id}}{ID of the user to edit or delete.}

\item{\code{name}}{Name to be inserted as user metadata.}

\item{\code{last_name}}{Last name to be inserted as user metadata.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-User-clone"></a>}}
\if{latex}{\out{\hypertarget{method-User-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
