% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_btns.R
\name{buttons}
\alias{buttons}
\alias{btn_add}
\alias{btn_trash}
\alias{btn_agregar}
\alias{btn_guardar}
\alias{btn_cancelar}
\alias{btn_eliminar}
\alias{btn_modificar}
\alias{btn_minus}
\alias{btn_refresh}
\alias{btn_expand}
\alias{btn_user_add}
\alias{btn_user_remove}
\alias{btn_user_edit}
\title{Pre styled pretty action buttons}
\usage{
btn_add(inputId, icon = NULL, size = "md", block = FALSE, no_outline = TRUE)

btn_trash(inputId, icon = NULL, size = "md", block = FALSE, no_outline = TRUE)

btn_agregar(
  inputId,
  icon = NULL,
  size = "md",
  block = FALSE,
  no_outline = TRUE
)

btn_guardar(
  inputId,
  icon = NULL,
  size = "md",
  block = FALSE,
  no_outline = TRUE
)

btn_cancelar(
  inputId,
  icon = NULL,
  size = "md",
  block = FALSE,
  no_outline = TRUE
)

btn_eliminar(
  inputId,
  icon = NULL,
  size = "md",
  block = FALSE,
  no_outline = TRUE
)

btn_modificar(
  inputId,
  icon = NULL,
  size = "md",
  block = FALSE,
  no_outline = TRUE
)

btn_minus(inputId, icon = NULL, size = "md", block = FALSE, no_outline = TRUE)

btn_refresh(
  inputId,
  icon = NULL,
  size = "md",
  block = FALSE,
  no_outline = TRUE
)

btn_expand(inputId, icon = NULL, size = "md", block = FALSE, no_outline = TRUE)

btn_user_add(
  inputId,
  icon = NULL,
  size = "md",
  block = FALSE,
  no_outline = TRUE
)

btn_user_remove(
  inputId,
  icon = NULL,
  size = "md",
  block = FALSE,
  no_outline = TRUE
)

btn_user_edit(
  inputId,
  icon = NULL,
  size = "md",
  block = FALSE,
  no_outline = TRUE
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{icon}{An optional icon to appear on the button.}

\item{size}{Size of the button : \code{xs},\code{sm}, \code{md}, \code{lg}.}

\item{block}{Logical, full width button.}

\item{no_outline}{Logical, don't show outline when navigating with
keyboard/interact using mouse or touch.}
}
\value{
HTML code for pretty action buttons
}
\description{
These functions have predefined styling and icons for labels.
The defaults are inherited from \code{\link[shinyWidgets:actionBttn]{shinyWidgets::actionBttn()}}.
}
\seealso{
\code{\link[shinyWidgets:actionBttn]{shinyWidgets::actionBttn()}}
}
