% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-DBManager.R
\name{DBManager}
\alias{DBManager}
\title{Clase R6 para manejar la base de datos}
\description{
Clase R6 para manejar la base de datos

Clase R6 para manejar la base de datos
}
\details{
Una base de datos puede ejecutar y obtener querys
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DBManager-new}{\code{DBManager$new()}}
\item \href{#method-DBManager-db_execute_statement}{\code{DBManager$db_execute_statement()}}
\item \href{#method-DBManager-db_get_query}{\code{DBManager$db_get_query()}}
\item \href{#method-DBManager-db_make_query}{\code{DBManager$db_make_query()}}
\item \href{#method-DBManager-clone}{\code{DBManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-DBManager-new}{}}}
\subsection{Method \code{new()}}{
Start the DB Manager
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBManager$new(use_tibble = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{use_tibble}}{Whether print the results of DbGetQuery() as a tibble or not. Default TRUE}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBManager-db_execute_statement"></a>}}
\if{latex}{\out{\hypertarget{method-DBManager-db_execute_statement}{}}}
\subsection{Method \code{db_execute_statement()}}{
Analog to  \code{DBI::dbExecute()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBManager$db_execute_statement(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Objects passed to glue::glue_sql() with the exception of the \code{.con} argument.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBManager-db_get_query"></a>}}
\if{latex}{\out{\hypertarget{method-DBManager-db_get_query}{}}}
\subsection{Method \code{db_get_query()}}{
Analog to  \code{DBI::dbGetQuery()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBManager$db_get_query(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Objects passed to glue::glue_sql() with the exception of the \code{.con} argument.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBManager-db_make_query"></a>}}
\if{latex}{\out{\hypertarget{method-DBManager-db_make_query}{}}}
\subsection{Method \code{db_make_query()}}{
Construct the query to be passed to \code{DBI::dbGetQuery()}. Useful for debugging and subquery construction.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBManager$db_make_query(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Objects passed to glue::glue_sql() with the exception of the \code{.con} argument.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DBManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
